Add IPC port to manifest
If the port for an application is not available, ql-tipc
immediately returns with an error without waiting for the port
unless that port is listed in the manifest. This adds the AVB
port to the manifest to fix a race between the bootloader
connecting to AVB and the app starting during boot.

Bug: 217580763
Change-Id: I77dcb9354cb3fbc389ee99193f3f2d398edff140